pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "apiguru-analytics"
version = "0.1.0"
description = "Analytics tools for OpenAPI specifications from APIs.guru"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Maciej" }
]
dependencies = [
    "polars",
    "numpy",
    "requests",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "mypy",
    "ruff",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
multi_line_output = 3 